Use BluetoothManager to get BluetoothAdapter
The `BluetoothAdapter.getDefaultAdapter()` method is deprecated in API level 31. This commit updates the code to use `BluetoothManager.getAdapter()` instead, which is the recommended way to get the `BluetoothAdapter` on API level 18 and higher.

This change ensures compatibility with newer Android versions and follows best practices for Bluetooth development. The `@RequiresApi(Build.VERSION_CODES.S)` annotation has been added to the `SelectBluetoothDevice` composable function to reflect the minimum API level required for this change.