Update Home.kt
The code changes include an update to the Home.kt file. The specific change is in the way the theme is toggled. The code now checks if ThemeController.isDarkMode is true, and if so, it sets the painter resource to R.drawable.light_mode_24px. Otherwise, it sets the painter resource to R.drawable.dark_mode_24px.